extend test-export
Previously, ~test-export~ just asserted that the exported JSON was valid JSON
and the $$data_type key of the top-level object was "org-document". This commit
introduces a JSON file representing the expected output of the test with some
keys whose values might change from one run to another (e.g. `ref`) omitted. To
compare this file with the actual output of the test, an optional argument was
added to the ~json-compare~ test helper that, when non-nil, causes the
comparison not to fail when keys in one object are missing from the other.